refactor(xlsx): extract table module from parser lib.rs
table.rs (~165 lines): tableStyles parsing, accent resolution
(ECMA-376 §18.8.40) and per-sheet table loading.

Final parser layout: lib.rs is now a 1452-line orchestration core
(public wasm entry points, parse_worksheet assembly, zip/theme/color
primitives, shared strings) plus six focused modules: types, styles,
chart, drawing, slicer, table. Down from the original monolithic 5804
lines (-75%). Verified behavior-invariant via VRT regression (65/65,
0.0% diff).
